Guard against null rootShadowNode in IntersectionObserverManager for initial notifications (#44528)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/44528

Changelog: [internal]

In the `IntersectionObserver` API we dispatch the initial notification in the `observe` method, but it might be possible that the surface has been removed from the registry by the time we execute that code.

This guards against that case to possibly fix a crash we're seeing in on the `IntersectionObserver` experiments.

Reviewed By: sammy-SC

Differential Revision: D57213994

fbshipit-source-id: 5cd1f16958949cc1ff64153e8012e6a59819d68a